+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PixPullSubscription.parse method
 - CreditSigner.resendToken method
 - Pdf and Payment methods to CreditNote resource
+- Update method to IssuingPurchase resource
+### Fixed
+- MerchantCategoryNumber type in IssuingPurchase resource, the API returns a string
 
 ## [1.2.0] - 2026-07-03
 ### Fixed